+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.1.30
+++++++
+* webapp: az webapp update supports httpsOnly
+* webapp/functionapp:  slot support for identity assign & identity show=======
+
 0.1.29
 ++++++
 * webapp/functionapp: author managed identity commands `identity assign/show`, and deprecate `assign-identity`