+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+2.0.29
+++++++
+Minor change
+
 2.0.28
 ++++++
 * vm/vmss create: support to attach unmanaged data disks and configure their caching modes